renderMessage cleanup
