+++ .npmrc
@@ -3,4 +3,3 @@ tag-version-prefix=""
 strict-peer-dependencies = false
 auto-install-peers = true
 lockfile = true
-shamefully-hoist=true
