+++ examples/simple-counter/src/CounterApp.tsx
@@ -9,18 +9,6 @@ type CounterAppProps = {
 export const CounterApp = ({ start = 0 }: CounterAppProps) => {
   const [count, setCount] = useState(start)
   const increment = () => setCount((count) => count + 1)
-  const timeoutRef = useRef<NodeJS.Timeout | null>(null)
-
-  useEffect(() => {
-    timeoutRef.current = setTimeout(() => {
-      console.log("tick " + count)
-      increment()
-    }, 2000)
-
-    return () => {
-      if (timeoutRef.current) clearInterval(timeoutRef.current)
-    }
-  }, [count])
 
   return (
     <>
