+++ packages/reaccord/src/Command.tsx
@@ -26,7 +26,7 @@ import type {
   UserContextMenuCommandInteraction,
 } from "discord.js"
 import type { MessageResponseOptions } from "./nodes"
-import type { RenderMessageFn } from "./renderer/renderMessage"
+import type { RenderMessageFn } from "./renderer/render"
 
 type CommandParamOptions<Required extends boolean = false> = {
   type: ApplicationCommandOptionType
