+++ packages/reaccord/src/renderer/render.tsx
@@ -1,13 +1,33 @@
 import { MessageProvider } from "../react/MessageContext"
+import { RootNode } from "../nodes"
 import { hostConfig } from "./hostConfig"
 import { isRootNode } from "../nodes/guards"
 import ReactReconciler from "react-reconciler"
-import type { ModalRootNode, RootNode } from "../nodes"
+import type { Client } from "../Client"
+import type {
+  InteractionRef,
+  MessageResponseOptions,
+  ModalRootNode,
+} from "../nodes"
+import type { Message } from "discord.js"
 
 export type RenderFn = (
   Code: () => JSX.Element,
   root: RootNode | ModalRootNode,
 ) => void
+
+export type RenderMessageFn = (
+  ref: InteractionRef,
+  Code: () => JSX.Element,
+) => Promise<Message>
+
+export const renderMessage =
+  (client: Client, rootOptions: MessageResponseOptions): RenderMessageFn =>
+  async (ref, Code) => {
+    const root = new RootNode(client, ref, rootOptions)
+    render(Code, root)
+    return root.updateMessage()
+  }
 
 const reactReconcilerInstance = ReactReconciler(hostConfig)
 reactReconcilerInstance.injectIntoDevTools({
